Initial commit for Zing.
Zing is going to be a simple wrapper around SQS to support laying off
database writes via Django to Amazon SQS.

This is highly experimental and not supposed to work the way I hope it
will. But in the end I hope to produce a high-performance worker library
to speedup database writes and layoff them from the request/response
cycle of modern web applications.